Make sure that the SkiaSharp Canvas is invalidated from the UI thread. Otherwise it doesn't work on iOS

+++ GradientControl/GradientControl/Controls/GradientStop.cs
@@ -4,11 +4,11 @@ namespace GradientControl.Controls
 {
 	public struct GradientStop
     {
-		public int Location { get; set; }
+		public float Location { get; set; }
 		public SKColor Color { get; set; }
 		public float Opacity { get; set; }
 
-	    public GradientStop(int location, SKColor color, float opacity)
+	    public GradientStop(float location, SKColor color, float opacity)
 	    {
 			Location = location;
 			Color = color;
